pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "markov-fut"
version = "0.5.0"
description = "Cadeias de Markov (StatsBomb) com P/S por equipe foco, reset por jogo e filtro robusto por eventos."
authors = [{ name = "Você" }]
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.24",
  "matplotlib>=3.8",
  "pygraphviz>=1.11",
  "networkx>=3.2",
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0"
]

[project.scripts]
markov-fut = "markov_fut.cli:app"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
